+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Add ability to request `AuthenticationInsight` when tokenizing a credit card, which can be used to make a decision about whether to perform 3D Secure verification
 * Include reference to Cardinal's docs for `uiCustomization` property on `ThreeDSecureRequest`.
 
 ## 3.6.0